Implemented the Login Page and Create Account page
