+++ Beef--it/App.xaml.cs
@@ -6,7 +6,7 @@
         {
             InitializeComponent();
 
-            MainPage = new NavigationPage(new LandingPage());
+            MainPage = new NavigationPage(new LoginPage());
         }
     }
 }
