+++ Beef--it/LoginPage/CreateAccountPage.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MauiApp1.CreateAccountPage"
+             x:Class="Beef__it.CreateAccountPage"
              Title="Create Account">
     <VerticalStackLayout
         Padding="30,0"
@@ -29,3 +29,4 @@
             Clicked="BackButtonClicked"/>
     </VerticalStackLayout>
 </ContentPage>
+
