--- Beef--it/LoginPage/HomePage.xaml.cs
@@ -1,9 +1,0 @@
-namespace MauiApp1;
-
-public partial class HomePage : ContentPage
-{
-	public HomePage()
-	{
-		InitializeComponent();
-	}
-}
