+++ Beef--it/LoginPage/LoginPage.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MauiApp1.MainPage"
-             Title="Login">
+             x:Class="Beef__it.LoginPage"
+             BackgroundColor="Black">
 
     <ScrollView>
         <VerticalStackLayout
