--- LoginPage/CreateAccountPage.xaml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MauiApp1.CreateAccountPage"
-             Title="Create Account">
-    <VerticalStackLayout
-        Padding="30,0"
-        Spacing="25">
-        <Label 
-            Text="Welcome to Beef-It"
-            VerticalOptions="Center" 
-            HorizontalOptions="Center"/>
-        <Label
-            Text="Create an account"
-            VerticalOptions="Center"
-            HorizontalOptions="Center"/>
-        <Entry
-            x:Name="newUsernameEntry"
-            Placeholder="Enter Username"/>
-        <Entry
-            x:Name="newPasswordEntry"
-            Placeholder="Enter Password"
-            IsPassword="True"/>
-        <Button
-            Text="Create"
-            Clicked="CreateAccountButtonClicked"/>
-        <Button
-            Text="Back to Sign-In"
-            Clicked="BackButtonClicked"/>
-    </VerticalStackLayout>
-</ContentPage>
