--- LoginPage/MainPage.xaml
@@ -1,41 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MauiApp1.MainPage"
-             Title="Login">
-
-    <ScrollView>
-        <VerticalStackLayout
-            Padding="30,0"
-            Spacing="25">
-
-            <Label
-                Text="Login" 
-                FontSize="24" 
-                HorizontalOptions="Center"/>
-
-            <Entry 
-                x:Name="UsernameEntry" 
-                Placeholder="Enter Username" />
-
-            <Entry
-                x:Name="PasswordEntry"
-                Placeholder="Enter Password"
-                IsPassword="True"/>
-
-            <Button
-                Text="Login"
-                Clicked="OnLoginClicked"/>
-
-            <Button 
-                Text="Create Account"
-                Clicked="OnCreateAccountClicked"/>
-
-            <Label
-                x:Name="ErrorMessage"
-                TextColor="Red"
-                IsVisible="False"/>
-        </VerticalStackLayout>
-    </ScrollView>
-    
-</ContentPage>
